% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_final.R
\name{alpha_sort_sp}
\alias{alpha_sort_sp}
\title{helper to sort species strings within each row}
\usage{
alpha_sort_sp(x)
}
\arguments{
\item{x}{list of strings (e.g. the species column)}
}
\value{
list of strings
}
\description{
helper to sort species strings within each row
}
