% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_step.R
\name{backward_step}
\alias{backward_step}
\title{A single backwards step}
\usage{
backward_step(
  remove_sp,
  comp_mat,
  initial_sp,
  ref_distmat,
  comp_dist,
  corr_method
)
}
\arguments{
\item{remove_sp}{set of variables to possibly exclude. Typically same as
\code{initial_sp}, unless \code{force_include} prohibits exclusion of some species}

\item{comp_mat}{variable community matrix}

\item{initial_sp}{set of variables already included}

\item{ref_distmat}{reference dissimilarity matrix}

\item{comp_dist}{\code{method} argument to use in \code{\link[vegan:vegdist]{vegan::vegdist()}} for dissimilarity}

\item{corr_method}{\code{method} argument to use in \code{\link[base:cor]{base::cor()}} for correlation}
}
\value{
named number, with the name of the best variable to remove and the
correlation value with it removed.
}
\description{
Cycles through each possible variable to remove and finds the best single
variable to remove from the \code{comp_mat}, as determined by increase in rho.
}
