% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvstep.R
\name{bvstep}
\alias{bvstep}
\title{Implements the bvstep algorithm from Clarke and Warwick}
\usage{
bvstep(
  ref_mat,
  comp_mat,
  ref_dist = "bray",
  comp_dist = "bray",
  rand_start = TRUE,
  nrand = round(ncol(ref_mat)/10),
  fixed_start = NULL,
  force_include = NULL,
  force_exclude = NULL,
  rho_threshold = 0.95,
  min_delta_rho = 0.001,
  corr_method = "kendall",
  selection_ref = "name"
)
}
\arguments{
\item{ref_mat}{the reference ('fixed') matrix that gets considered in whole}

\item{comp_mat}{the 'variable' matrix that gets forward/backward chopped}

\item{ref_dist}{distance metric for \code{ref_mat} (see \code{method} in \code{\link[vegan:vegdist]{vegan::vegdist()}})}

\item{comp_dist}{distance metric for \code{comp_mat} (see \code{method} in \code{\link[vegan:vegdist]{vegan::vegdist()}})}

\item{rand_start}{logical, default \code{TRUE}.
\itemize{
\item \code{TRUE}, start with \code{nrand} randomly-selected columns from \code{comp_mat} (plus any in \code{force_include}). Useful with \code{\link[=bv_multi]{bv_multi()}} to avoid local optima.
\item \code{FALSE}, start with the best available single column of \code{comp_mat} if \code{force_include = NULL} or \code{force_include}.
}}

\item{nrand}{number of columns to choose for a random start. Defaults to 10\%. Should not be too large, or it is hard to drop uninformative. Only used if \code{rand_start = TRUE}}

\item{fixed_start}{\code{NULL} (default) or character of columns to start with. Allows forcing starts with set subset of the data.}

\item{force_include}{\code{NULL} (default) or character of columns to always include- these are there at the start and never get dropped. If \code{rand_start = TRUE}, random columns are in addition to these.}

\item{force_exclude}{\code{NULL} (default) or character of columns to always exclude- these are just dropped right at the start.}

\item{rho_threshold}{Default 0.95. Threshold at which to cut off the process. If this is crossed at a forward step, one last backward step is taken to check if a smaller set still meets the condition.}

\item{min_delta_rho}{Default 0.001. Cutoff to stop the process if rho is not increasing much.}

\item{corr_method}{character, default \code{kendall}. This is the \code{method} argument of \code{\link[=cor]{cor()}}. Clarke and Warwick 1998 suggests kendall because we need a rank correlation and spearman doesn't handle ties.}

\item{selection_ref}{character, \code{'name'} (default)- return the variable names in each step. \code{'index'} returns their column index in \code{comp_mat}}
}
\value{
a data frame (tibble) with the iteration, whether it was forward/backward, correlation, and set of column names selected.
}
\description{
If we start with \code{rand_start} and \code{nrand} is more than 1, this starts with backward steps. If there is only a single column start (e.g. \code{rand_start = FALSE}) or only \code{force_include}, starts with forward selection. Note that though there is only ever a single forward step before backward, backward steps do not get saved into the dataframe if they do not increase rho, and forward steps do not get entered if they increase rho by less than \code{min_delta_rho} (and the steps then terminate).
}
\examples{
require(vegan)
data(varespec)
bvout <- bvstep(
  ref_mat = varespec, comp_mat = varespec,
  ref_dist = "bray", comp_dist = "bray",
  rand_start = TRUE, nrand = 5
)
}
