% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcorr.R
\name{distcorr}
\alias{distcorr}
\title{Find the community distance for a variable community matrix and then the rank
correlation with a reference dissimilarity matrix}
\usage{
distcorr(comp_mat, ref_distmat, comp_dist, corr_method)
}
\arguments{
\item{comp_mat}{variable community matrix}

\item{ref_distmat}{reference dissimilarity matrix}

\item{comp_dist}{\code{method} argument to use in \code{\link[vegan:vegdist]{vegan::vegdist()}} for dissimilarity}

\item{corr_method}{\code{method} argument to use in \code{\link[stats:cor]{stats::cor()}} for correlation}
}
\value{
correlation value
}
\description{
Find the community distance for a variable community matrix and then the rank
correlation with a reference dissimilarity matrix
}
