% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_final.R
\name{extract_final}
\alias{extract_final}
\title{Extract the final set from a dataframe or list of bvstep outputs}
\usage{
extract_final(bvout)
}
\arguments{
\item{bvout}{bvstep outputs, either dataframe or list}
}
\value{
dataframe
}
\description{
Extract the final set from a dataframe or list of bvstep outputs
}
