% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_names.R
\name{extract_names}
\alias{extract_names}
\title{Get names as a character vector instead of a single string}
\usage{
extract_names(bvout, step = "last")
}
\arguments{
\item{bvout}{output list or dataframe of a bvstep}

\item{step}{step (numeric) to extract. Default 'last' uses \code{nrow(bvout)}.}
}
\value{
character vector
}
\description{
To match PRIMER, the names are included as a single string with comma
separation. For further use, though, we need to parse that back to a
character vector
}
