% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_step.R
\name{forward_step}
\alias{forward_step}
\title{A single forward step}
\usage{
forward_step(comp_mat, initial_sp, ref_distmat, comp_dist, corr_method)
}
\arguments{
\item{comp_mat}{variable community matrix}

\item{initial_sp}{set of variables already included}

\item{ref_distmat}{reference dissimilarity matrix}

\item{comp_dist}{\code{method} argument to use in \code{\link[vegan:vegdist]{vegan::vegdist()}} for dissimilarity}

\item{corr_method}{\code{method} argument to use in \code{\link[stats:cor]{stats::cor()}} for correlation}
}
\value{
named number, with the name of the best variable to add and the
correlation value with it included.
}
\description{
Cycles through each possible variable to add and finds the best single
variable to add to the \code{comp_mat}, as determined by increase in rho.
}
