% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{null_transformer}
\alias{null_transformer}
\title{Let glue handle NULL and return 'NULL'
from https://github.com/tidyverse/glue/issues/100}
\usage{
null_transformer(str = "NULL")
}
\arguments{
\item{str}{string to return for NULL}
}
\value{
string for glue where there are NULLs
}
\description{
Let glue handle NULL and return 'NULL'
from https://github.com/tidyverse/glue/issues/100
}
