% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peeler-package.R
\docType{package}
\name{peeler-package}
\alias{peeler}
\alias{peeler-package}
\title{peeler: Implements peeling community datasets using sequential bvstep algorithms, as in Clarke and Warwick 1998}
\description{
This package provides implementations of the bvstep forward/backward algorithm to find the best subset of the data that matches the full community. It further provides functions to iterate that algorithm over some number of random starts to assess consistency. The `peel` function then uses these to iteratively find the best subset and remove it.
}
\author{
\strong{Maintainer}: Galen Holt \email{g.holt@deakin.edu.au} (\href{https://orcid.org/0000-0002-7455-9275}{ORCID})

}
\keyword{internal}
