% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_step.R
\name{single_backward}
\alias{single_backward}
\title{Backward step for a single variable}
\usage{
single_backward(
  cut_sp,
  comp_mat,
  initial_sp,
  ref_distmat,
  comp_dist,
  corr_method
)
}
\arguments{
\item{cut_sp}{single variable to remove}

\item{comp_mat}{variable community matrix}

\item{initial_sp}{set of variables already included}

\item{ref_distmat}{reference dissimilarity matrix}

\item{comp_dist}{\code{method} argument to use in \code{\link[vegan:vegdist]{vegan::vegdist()}} for dissimilarity}

\item{corr_method}{\code{method} argument to use in \code{\link[stats:cor]{stats::cor()}} for correlation}
}
\value{
named num, correlation, with name being the removed species
}
\description{
This is run for each possible removal in turn.
}
