% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_step.R
\name{single_forward}
\alias{single_forward}
\title{Forward step for a single variable}
\usage{
single_forward(
  add_sp,
  comp_mat,
  initial_sp,
  ref_distmat,
  comp_dist,
  corr_method
)
}
\arguments{
\item{add_sp}{single variable to add}

\item{comp_mat}{variable community matrix}

\item{initial_sp}{set of variables already included}

\item{ref_distmat}{reference dissimilarity matrix}

\item{comp_dist}{\code{method} argument to use in \code{\link[vegan:vegdist]{vegan::vegdist()}} for dissimilarity}

\item{corr_method}{\code{method} argument to use in \code{\link[stats:cor]{stats::cor()}} for correlation}
}
\value{
named number, correlation with the name being the added species
}
\description{
This is run for each possible addition in turn.
}
